Remove function which didn't do its job anyway.
The function never worked, because the form data for the ajax request is
already assembled at this point in time. So modifying the input values
afterwards makes no sense.
The server already handles this situation. So instead of fixing that by
overwriting the assembled ajax form date, I just removed the function.
Less code is best code :)

This fixes OlliL/moneyjinn-server#21